+++ install/setup-fixtures.php
@@ -1,6 +1,8 @@
 #!/usr/bin/env php
 <?php
 
+use AKlump\FixtureFramework\Runtime\RunOptions;
+
 $vendor_dir = __DIR__ . '/../vendor';
 require_once $vendor_dir . '/autoload.php';
 
@@ -25,14 +27,19 @@ catch (Exception $e) {
 }
 
 try {
-  $options = [
+  $run_options_validator = new \AKlump\FixtureFramework\Runtime\RunOptionsValidator();
+  $run_options = new RunOptions([
       'env' => 'test',
       'url' => 'https://website.com/',
       'drush' => 'lando nxdb_drush',
-  ];
+      $run_options_validator,
+  ]);
+
   $run_context_validator = new \AKlump\FixtureFramework\Runtime\RunContextValidator();
-  $instantiator = new \AKlump\FixtureFramework\Runtime\FixtureInstantiator($options, $run_context_validator);
+  $instantiator = new \AKlump\FixtureFramework\Runtime\FixtureInstantiator($run_options, $run_context_validator);
+
   $fixtures = (new \AKlump\FixtureFramework\Runtime\FixtureCollectionBuilder($instantiator))($definitions);
+
   $runner = new \AKlump\FixtureFramework\Runtime\FixtureRunner($fixtures);
   $runner->run($silent);
 }